requirements.txt
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
pytoshop>=1.5.0
psd-tools>=1.9.0
torch>=2.0.0
